fix build
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@5854 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr_menu.c
@@ -279,7 +279,7 @@ static switch_status_t play_or_say(switch_core_session_t *session, switch_ivr_me
 		if (need) {
 			menu->ptr += strlen(menu->buf);
 			if (strlen(menu->buf) < need) {
-				switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_DEBUG, "waiting for %u digits\n", need);
+				switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_DEBUG, "waiting for %u digits\n", (uint32_t)need);
 				status = switch_ivr_collect_digits_count(session, menu->ptr, menu->inlen - strlen(menu->buf), need, "#", &terminator, menu->timeout);
 			}
 			switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_DEBUG, "digits '%s'\n", menu->buf);
